backend/requirements.txt
# Core web framework
fastapi==0.128.0
uvicorn==0.40.0
python-dotenv==1.2.1

# Database
SQLAlchemy==2.0.45

# Data validation
pydantic==2.12.5

# Supabase client
supabase

# LangGraph and LangChain
langgraph
langchain-core

# HTTP client
httpx
